Exception handling - delete

+++ src/main/java/com/mydomini/courseJava/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.mydomini.courseJava.entities.User;
 import com.mydomini.courseJava.repositories.UserRepository;
+import com.mydomini.courseJava.services.exceptions.DatabaseException;
 import com.mydomini.courseJava.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -21,9 +23,9 @@ public class UserService {
 		return repository.findAll();
 	}
 	
-	public User findById(Long Id) {
-		Optional<User> obj = repository.findById(Id); 
-		return obj.orElseThrow(() -> new ResourceNotFoundException(Id));
+	public User findById(Long id) {
+		Optional<User> obj = repository.findById(id); 
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 	
 	public User insert(User obj) {
@@ -31,7 +33,13 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
+	    if (!repository.existsById(id)) {
+	        throw new ResourceNotFoundException(id);
+	    }try {
 		repository.deleteById(id);
+		}catch(DataIntegrityViolationException  e){
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User obj) {
